app/requirements.txt
streamlit==1.36.0
psycopg2-binary==2.9.9
pandas==2.2.2
